resp

+++ app/Http/Controllers/TwitterController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Http\Api\CallTwitterApi;
 use App\Models\Tweet;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 class TwitterController extends Controller
 {
@@ -14,16 +15,11 @@ class TwitterController extends Controller
         $keyword = $request->input('searchWord');
         $id = $request->user()->id;
         if(!empty($keyword)){
-            //dd($keyword);
             $searchResults = $connection->searchTweets($keyword);
         }
         else {
             $searchResults = $connection->searchTweets("テスト");
         }
-        //$array = array();
-        //foreach($searchResults as $searchResults){
-            //$array[] = array($connection->statusesOembed($searchResults->id));
-        //}
         return view('twitter', ['twitter' => $searchResults])->with('id', $id);
     }
 
@@ -32,7 +28,6 @@ class TwitterController extends Controller
         $tweets = $requests->all()['check'];
         foreach($tweets as $number => $value) {
             $tweet = new Tweet;
-            //dd($requests->user()->id);
             $tweet->user_id = $requests->user()->id;
             $tweet->content = $value;
             $tweet->save();
@@ -42,20 +37,27 @@ class TwitterController extends Controller
 
     public function keep(Request $request)
     {
-        //dd($request->route('id'));
         $tweets = Tweet::where('user_id', '=', $request->route('id'))->get();
         $id = $request->user()->id;
-        //dd($request->user()->id);
+        if($request->route('id') != $id)
+        {
+            throw new AccessDeniedHttpException();
+        }
         return view('keep', ['id' => $id])->with('tweets',$tweets);
     }
 
     public function delete(Request $requests)
     {
         $tweets = $requests->all()['check'];
+        $id = $requests->user()->id;
+        if($requests->route('id') != $id)
+        {
+            throw new AccessDeniedHttpException();
+        }
         foreach($tweets as $number => $id) {
             $tweetData = Tweet::where('id',$id)->firstOrFail();
             $tweetData->delete();
         }
-        return redirect()->route('tweet.keep')->with('feedback.success',"削除に成功しました");
+        return redirect()->route('tweet.keep', ['id' => $requests->user()->id])->with('feedback.success',"削除に成功しました");
     }
 }
